(PC-2113): Move count_activated_users() to user repository
